extract indexer into a separate class
